Move mentions into separate subfolder for clearity
