+++ app/lib/common/toolkit/html_editor/html_editor.dart
@@ -4,14 +4,13 @@ import 'dart:math';
 import 'package:acter/common/extensions/options.dart';
 import 'package:acter/common/toolkit/buttons/primary_action_button.dart';
 import 'package:acter/common/toolkit/buttons/user_chip.dart';
-import 'package:acter/common/toolkit/html_editor/services/commands/backspace_for_mentions.dart';
-import 'package:acter/common/toolkit/html_editor/services/mention_detection.dart';
+import 'package:acter/common/toolkit/html_editor/mentions/commands/backspace_for_mentions.dart';
+import 'package:acter/common/toolkit/html_editor/mentions/mention_detection.dart';
 import 'package:acter/config/constants.dart';
-import 'package:acter/common/toolkit/html_editor/models/mention_attributes.dart';
-import 'package:acter/common/toolkit/html_editor/models/mention_type.dart';
+import 'package:acter/common/toolkit/html_editor/mentions/models/mention_attributes.dart';
+import 'package:acter/common/toolkit/html_editor/mentions/models/mention_type.dart';
 import 'package:acter/common/toolkit/html_editor/services/constants.dart';
-import 'package:acter/common/toolkit/html_editor/services/mention_shortcuts.dart';
-import 'package:acter/features/deep_linking/parse_acter_uri.dart';
+import 'package:acter/common/toolkit/html_editor/mentions/mention_shortcuts.dart';
 import 'package:acter/features/deep_linking/types.dart';
 import 'package:acter/features/deep_linking/widgets/inline_item_preview.dart';
 import 'package:acter/features/room/widgets/room_chip.dart';
