+++ app/lib/common/toolkit/html_editor/mentions/mention_detection.dart
